asmparsertest: Cast isspace input to int to silence compiler warning.
Fixes this GCC warning on Cygwin.
asmparsertest.c: In function ‘compile’:
asmparsertest.c:136:4: warning: array subscript has type ‘char’

+++ tests/asmparsertest/asmparsertest.c
@@ -133,7 +133,7 @@ compile(const char *filename, GLenum target, int use_ARB)
 
 			ptr += strlen("# REQUIRE ");
 
-			for (i = 0; !isspace(ptr[i]) && (ptr[i] != '\0'); i++) {
+			for (i = 0; !isspace((int) ptr[i]) && (ptr[i] != '\0'); i++) {
 				extension[i] = ptr[i];
 			}
